Remove useAutoAsync
The wrapping of other hook libraries has been a maintenence thorn due to
frequent breaking changes. The breaking changes didn't often impact this
library, but the peer-dep constraint made it difficult to update this
library nonetheless.

This hook was the last hold-out and `react-query` or similar is
typically a better solution. If nothing else, the "auto" portion is
relatively easily achievable with an effect.